+++ gcs/src/test/java/com/google/cloud/hadoop/fs/gcs/GoogleHadoopFileSystemConfigurationTest.java
@@ -68,9 +68,6 @@ public class GoogleHadoopFileSystemConfigurationTest {
           put("fs.gs.block.size", 64 * 1024 * 1024L);
           put("fs.gs.bucket.delete.enable", false);
           put("fs.gs.checksum.type", GcsFileChecksumType.NONE);
-          put("fs.gs.cooperative.locking.enable", false);
-          put("fs.gs.cooperative.locking.expiration.timeout.ms", 120_000L);
-          put("fs.gs.cooperative.locking.max.concurrent.operations", 20);
           put("fs.gs.copy.with.rewrite.enable", true);
           put("fs.gs.create.items.conflict.check.enable", true);
           put("fs.gs.delegation.token.binding", null);